Fix probabilities do not sum to 1

+++ src/qibo/backends/numpy.py
@@ -63,7 +63,7 @@ class NumpyBackend(abstract.AbstractBackend):
         if isinstance(dtype, str):
             dtype = self.dtypes(dtype)
         if isinstance(x, self.backend.ndarray):
-            return x.astype(dtype)
+            return x.astype(dtype, copy=False)
         return self.backend.array(x, dtype=dtype)
 
     def diag(self, x, dtype='DTYPECPX'):
